ZerePy/pyproject.toml
[tool.poetry]
name = "ZerePy"
version = "0.1.0"
description = "A launch-pad for AI agents"
authors = [
    "Ayoub Ed <ayoub@blorm.xyz>",
    "Efrain <ef@blorm.xyz>",
    "Max <max@blorm.xyz>",
    "Dev Arogundade <devarogundade@gmail.com>",
]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
openai = "^1.57.2"
prompt-toolkit = "^3.0.48"
anthropic = "^0.42.0"
farcaster = "^0.7.11"
goat-sdk = "^0.1.1"
goat-sdk-wallet-web3 = "^0.1.1"
web3 = "^6.20.3"
goat-sdk-plugin-erc20 = "^0.1.0"
goat-sdk-plugin-coingecko = "^0.1.1"
solders = "^0.21.0,<0.24.0"
solana = "^0.35.0"
aiohttp = "^3.11.11"
requests = "2.32.3"
jupiter-python-sdk = "^0.0.2.0"
allora-sdk = "^0.1.0"
requests-oauthlib = "^1.3.1"
together = "^1.3.14"
fastapi = { version = "^0.109.0", optional = true }
uvicorn = { version = "^0.27.0", optional = true }
pymongo = "^4.11.1"
motor = "^3.7.0"

[tool.poetry.extras]
server = ["fastapi", "uvicorn", "requests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
